major code refactor to new API and naming
This commit was a big one. The UserKey is completely gone now in favour of a state machine Nym. Nym can be a deterministic Nym<Initial> which can be used to accept credentials for retrieval from storage in a future date. Or, a Nym<Randomized> can be used to get issued, offer, and prove Creds anonymously. This should be more ergonomic yet no-foot-gun safe way of using the API